requirements.txt
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
requests==2.27.1
Jinja2==3.0.3
python-dateutil==2.8.2
